+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Made Insurance Policy chest spawn at the start of any run, even if the shortcut elevator is used
 - Fixed Gun Powderer being able to spawn more than 5 ammo boxes
+- Fixed dummy items from appearing in the inventory on the victory / death screen
 
 ## 1.15.0 (2024-07-05)